problem:
please find attached two sets of images (two different patients). in
each case, there are two images, one from the left and another from the
right breast.  one side has cancer and the other side is healthy.
case181 and 182 are the same patient. and case5 and 5 are another
patient. u can use image registration or any other technique of your
choice to locate the site of cancer. if u are not familiar with
registration, please focus on the machine learning problem below.
i have also attached a multi-variate time series data. this was a
three-class classification problem. the goal column has three values 0,
1, and 2.  zero means there was no event in that time interval.  to make
it easier treat both 1 and 2 cases as the same (so it would be a
two-class classification). the other columns are the variables listed in
time. for example rows 2 to 90 represent one time series, etc.